New group screen

+++ Q-municate/Categories/UIImage+fixOrientation.m
@@ -63,7 +63,7 @@
     
     // Now we draw the underlying CGImage into a new context, applying the transform
     // calculated above.
-    CGContextRef ctx = CGBitmapContextCreate(NULL, self.size.width, self.size.height,
+    CGContextRef ctx = CGBitmapContextCreate(NULL, (size_t)self.size.width, (size_t)self.size.height,
                                              CGImageGetBitsPerComponent(self.CGImage), 0,
                                              CGImageGetColorSpace(self.CGImage),
                                              CGImageGetBitmapInfo(self.CGImage));
